pyproject.toml
[project]
name = "taskflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "F3445038 Francisco Manoel Da Silva Filho", email = "francisco.filho@bb.com.br" }
]
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.44",
    "google-genai>=1.20.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]

[project.scripts]
taskflow = "taskflow:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
